refactor: move agoric proposals to config/agoric/index.tsx
